requirements.txt
Flask==3.0.0
pymongo==4.6.1
python-telegram-bot==20.7
gunicorn==21.2.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
anthropic==0.40.0
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
reportlab==4.0.9
PyPDF2==3.0.1
cryptography==41.0.7
python-dateutil==2.8.2
